[Feat ✏️] 내 루틴에 속한 운동정보 가져오기 hook 분리
